refactor(snapshot): isolate Windows ACL policy
